refactor(parent-grounding-start-screen): ParentGroundingStack
