+++ app/screens/DummyQuests/DummyQuests.tsx
@@ -2,7 +2,6 @@ import { values } from 'lodash';
 import React, { useCallback } from 'react';
 import { SafeAreaView, TouchableOpacity } from 'react-native';
 import { FlatList } from 'react-native-gesture-handler';
-import { IMAGES } from '../../assets';
 
 import { ExtendedText, ExtendedButton } from '../../components';
 import { useAppDispatch, useAppSelector, useMount } from '../../hooks';
@@ -74,27 +73,6 @@ export const DummyQuestsScreen: React.FC<IDummyQuestsScreenProps> = ({
     navigation.navigate('RecognitionStack');
   }, [navigation]);
 
-  const goToParentGroundingStack = useCallback(() => {
-    navigation.navigate('ParentGroundingStack', {
-      screen: 'GroundingStart',
-      params: {
-        data: {
-          image: IMAGES.LOGO,
-          titleKey: 'screens.parent_grounding_exercise.grounding_start.title',
-          subtitleKey:
-            'screens.parent_grounding_exercise.grounding_start.subtitle',
-          aboutTitleKey:
-            'screens.parent_grounding_exercise.grounding_start.about_title',
-          aboutSubtitleKey:
-            'screens.parent_grounding_exercise.grounding_start.about_subtitle',
-          topBackground: IMAGES.WHITE_BACK_ARROW,
-          bottomBackground: IMAGES.WHITE_BACK_ARROW,
-          nextRouteName: 'ComingSoon',
-        },
-      },
-    });
-  }, [navigation]);
-
   return (
     <SafeAreaView style={styles.container}>
       <ExtendedText preset="large-title" style={styles.buttonsTitle}>
@@ -108,11 +86,6 @@ export const DummyQuestsScreen: React.FC<IDummyQuestsScreenProps> = ({
       <ExtendedButton
         onPress={goToRecognitionStack}
         title="To recognition"
-        style={styles.button}
-      />
-      <ExtendedButton
-        onPress={goToParentGroundingStack}
-        title="To parent grounding"
         style={styles.button}
       />
       <ExtendedText preset="large-title" style={styles.questsTitle}>
