--- app/screens/Grounding/General/GroundingStart/index.ts
@@ -1,2 +1,0 @@
-export { GroundingStartScreen } from './GroundingStart';
-export type { IGroundingStartRoute } from './GroundingStart.types';
