--- app/screens/Grounding/General/index.ts
@@ -1,1 +1,0 @@
-export * from './GroundingStart';
