+++ app/screens/Initial/InitialScreen.tsx
@@ -56,7 +56,7 @@ export const InitialScreen: React.FC<IInitialScreenProps> = ({
       isStartLoading
     ) {
       if (user) {
-        navigation.replace('GardenStack');
+        navigation.replace('ParentGroundingStack');
       } else {
         navigation.replace('ParentsOnboardingStack');
       }
