+++ app/screens/ParentGrounding/GroundingStart/index.ts
@@ -1,0 +1,1 @@
+export { GroundingStartScreen } from './GroundingStart';
