+++ app/screens/ParentGrounding/index.ts
@@ -1,1 +1,2 @@
 export * from './GroundingAcknowledgement';
+export * from './GroundingStart';
