+++ app/screens/index.ts
@@ -1,7 +1,7 @@
 export * from './Initial';
 export * from './DummyQuests';
 export * from './Garden';
-export * from './Grounding';
+export * from './ParentGrounding';
 export * from './Menu';
 export * from './MixingElixir';
 export * from './Onboarding';
